requirements.txt
certifi==2021.10.8
chardet==3.0.4
Deprecated==1.2.13
idna==2.10
PyGithub==1.51
PyJWT==2.4.0
python-dotenv==0.14.0
requests==2.24.0
urllib3==1.25.11
wrapt==1.14.1
